[feat] 게시글 삭제, 수정 시 Post의 PK값 이용으로 변경
- 해당 게시글을 찾을 때 Post의 PK값으로 찾기로 변경
- 게시글 수정 시, 변경 전 제목 입력 삭제
- 게시글 수정 입력 폼에 이전 제목 입력 -> PostID값 입력
- 게시글 수정, 삭제 시 로그인한 사용자와 작성자가 동일 사용자인지 확인 후 다르다면 에러 발생

+++ src/main/java/org/board/exercise_board/Post/application/RemoveApplication.java
@@ -1,10 +1,12 @@
 package org.board.exercise_board.Post.application;
 
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.board.exercise_board.Post.domain.model.Post;
+import org.board.exercise_board.Post.exception.PostCustomException;
+import org.board.exercise_board.Post.exception.PostErrorCode;
 import org.board.exercise_board.Post.service.PostService;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -12,11 +14,15 @@ public class RemoveApplication {
 
   private final PostService postService;
 
-  @Transactional
-  public String removePost(String writerId, String subject) {
-    Post post = postService.findPosts(writerId, subject);
-    postService.removePost(post);
-    return "게시물 삭제가 완료되었습니다.";
+  public String removePost(String writerId, Long postId) {
+    Post post = postService.findPost(postId);
+
+    // 로그인한 사용자가 작성자인지 확인
+    if(!Objects.equals(writerId, post.getUser().getLoginId())) {
+      throw new PostCustomException(PostErrorCode.NOT_HAVE_RIGHT);
+    }
+
+    return postService.removePost(post);
   }
 
 }
